fix: correct spelling errors, improve error handling, and enhance functionality
- Fixed spelling error in BarcodeFormat enum from UNSUPORTED_FORMAT to UNSUPPORTED_FORMAT.
- Updated error messages for unsupported formats in Display implementation.
- Improved BinaryBitmap display logic to handle uninitialized states.
- Changed BitArray initialization method from with_capacity to with_size.
- Enhanced BitArray seek method to handle out-of-bounds errors more gracefully.
- Fixed cropping logic in BitMatrix to correctly adjust coordinates.
- Updated BitSource read method to correctly calculate available bytes.
- Added UTF16LE character set support in CharacterSet.
- Corrected BitMatrix cursor logic to ensure proper edge stepping.
- Improved error handling in DMRegressionLine average calculation.
- Enhanced edge tracing logic to ensure proper corner tracing.
- Fixed FastEdgeToEdgeCounter index calculation to prevent negative indices.
- Updated Matrix struct to ensure proper data length validation.
- Improved error handling in ReedSolomonEncoder for EC bytes.
- Corrected ECIInput interface to use usize instead of u32 for index.
- Updated Scan component to simplify QR decoding logic.
- Refined rxingWorkerUtils to ensure consistent return types.

+++ rust/rxing-vendored/src/common/eci_input.rs
@@ -87,7 +87,7 @@ pub trait ECIInput: Display {
      *          if the {@code index} argument is negative or not less than
      *          {@code length()}
      */
-    fn isECI(&self, index: u32) -> Result<bool>;
+    fn isECI(&self, index: usize) -> Result<bool>;
 
     /**
      * Returns the {@code int} ECI value at the specified index.  An index ranges from zero
